Trust STS IDP connection when the url is localhost
During SSO login, Console contacts MinIO server to generate new
temporary credentials. When TLS is enabled, setting up a correct
TLS certificate is something that needs to be done correctly by
the user. However, recently, we started to skip the TLS verification
when Console talks to MinIO server using a loopback address, but we
missed the case of Console generating temporary credentials in case of
IDP. This commit will get the configured MinIO server url to decide
if the STS call needs to skip the TLS verification or not.