fix: handling of no inline policy for service acc.
When no inline policy is specified for a service account, stop sending
`null` as the policy JSON value to the create service account API
endpoint.

Also, skip the unnecessary console side validation of the policy JSON as
the create-service-account API already does this - no client-server
round-trip is saved by doing it. Console was also re-encoding the user
provided JSON due to this - now this is prevented as well.

This is part of the remaining work mentioned in
https://github.com/minio/minio/pull/18945