fix health report upload error
After moving the processing of uploaded health report to background,
SUBNET now doesn't return the cluster URL in response of health upload.

This was resulting in error being shown on the UI even when upload to
SUBNET was succesful.

Fix this by not expecting the cluster URL in response, and instead
linking to the deployments page URL (/health) after succesful upload.

Also start using exposed functions from mc instead of duplicating the
logic in console.